Move settings components
