--- src/components/plugin/index.js
@@ -1,1 +1,0 @@
-export * from './pane'
