+++ src/components/project/settings.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { ipcRenderer as ipc } from 'electron'
 import { arrayOf, func, object, string } from 'prop-types'
 import { Form, FormField, FormToggle, FormToggleGroup } from '../form.js'
-import { ProjectTypeField } from './project-type-field.js'
+import { ProjectTypeField } from './type-field.js'
 import { notify } from '../../dialog.js'
 import { BASES, TYPES } from '../../common/project.js'
 import { convert } from '../../slices/project-files.js'
