+++ src/components/settings/locale.js
