+++ src/components/vocab/browser.js
